+++ sdk/pixUser/index.js
@@ -1,4 +1,6 @@
 const pixUser = require('./pixUser.js');
 
+exports.statistics = require('./statistics');
+exports.Statistics = exports.statistics.Statistics;
 exports.get = pixUser.get;
 exports.PixUser = pixUser.PixUser;