+++ sdk/pixUser/statistics/index.js
@@ -1,0 +1,2 @@
+const statistics = require('./statistics.js');
+exports.Statistics = statistics.Statistics;